pyproject.toml
[tool.poetry]
name = "reviews-kaz-classification"
version = "0.1.0"
description = "package for training classfication model"
authors = ["alimuteshov <alimuteshov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.1.1"
torch = "^2.0.1"
transformers = "^4.33.2"
numpy = "^1.26.0"
scikit-learn = "^1.3.1"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
dvc = {extras = ["gdrive"], version = "^3.30.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
